dom/media/tests/mochitest/test_getUserMedia_constraints_mobile.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=882145
-->
<head>
  <meta charset="utf-8">
  <title>Test mozGetUserMedia Constraints</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript" src="constraints.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=882145">Test mozGetUserMedia Constraints (mobile)</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript">
/**
  See constraints.js for testConstraints() and common_tests.
  TODO(jib): Merge desktop and mobile version of these tests again (Bug 997365)
*/
var mobile_tests = [
  { message: "legacy facingMode overconstrains video (mobile)",
    constraints: { video: { mandatory: { facingMode:'left' } }, fake: true },
    error: "NO_DEVICES_FOUND" },
];

runTest(function () {
  testConstraints(common_tests.concat(mobile_tests));
});


</script>
</pre>
</body>
</html>
